NODE-820: Move contract_hash to cl_node/contract_hash.py to keep all library code in the cl_node lib/framework.
